+++ server/src/index.ts
@@ -9,6 +9,7 @@ import { enhancedTools } from './tools/enhanced_tools.js';
 import { scriptResourceTools } from './tools/script_resource_tools.js';
 import { debuggerTools } from './tools/debugger_tools.js';
 import { projectTools } from './tools/project_tools.js';
+import { inputTools } from './tools/input_tools.js';
 import { getGodotConnection } from './utils/godot_connection.js';
 
 // Import resources
@@ -63,7 +64,8 @@ async function main() {
     ...enhancedTools,
     ...projectTools,
     ...scriptResourceTools,
-    ...debuggerTools
+    ...debuggerTools,
+    ...inputTools
   ];
 
   allTools.forEach(tool => {